+++ src/main/java/de/markusbordihn/playercompanions/entity/type/collector/CollectorFeatures.java
@@ -41,7 +41,6 @@ import de.markusbordihn.playercompanions.entity.PlayerCompanionsFeatures;
 public class CollectorFeatures extends PlayerCompanionsFeatures {
 
   private static final CommonConfig.Config COMMON = CommonConfig.COMMON;
-  private static int collectorTypeRadius = COMMON.collectorTypeRadius.get();
 
   private static final short COLLECT_TICK = 20 * 3;
 
@@ -51,10 +50,9 @@ public class CollectorFeatures extends PlayerCompanionsFeatures {
 
   @SubscribeEvent
   public static void handleServerAboutToStartEvent(ServerAboutToStartEvent event) {
-    collectorTypeRadius = COMMON.collectorTypeRadius.get();
-    if (collectorTypeRadius > 0) {
+    if (COMMON.collectorTypeRadius.get() > 0) {
       log.info("{} Collector will automatically collect items in a {} block radius.",
-          Constants.LOG_ICON, collectorTypeRadius);
+          Constants.LOG_ICON, COMMON.collectorTypeRadius.get());
     } else {
       log.info("{} Collector will not automatically collect items!", Constants.LOG_ICON);
     }
@@ -62,9 +60,9 @@ public class CollectorFeatures extends PlayerCompanionsFeatures {
 
   private void collectorTick() {
     // Automatic collect items in the defined radius
-    if (!this.level.isClientSide && collectorTypeRadius > 0 && ticker++ >= COLLECT_TICK) {
+    if (!this.level.isClientSide && COMMON.collectorTypeRadius.get() > 0 && ticker++ >= COLLECT_TICK) {
       List<ItemEntity> itemEntities = this.level.getEntities(EntityType.ITEM,
-          new AABB(playerCompanionEntity.blockPosition()).inflate(collectorTypeRadius),
+          new AABB(playerCompanionEntity.blockPosition()).inflate(COMMON.collectorTypeRadius.get()),
           entity -> true);
       if (!itemEntities.isEmpty()) {
         PlayerCompanionData companionData = playerCompanionEntity.getData();